requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
spacy>=3.7.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
pytest>=8.0.0
httpx>=0.27.0        # for FastAPI TestClient
google-genai>=1.0.0
python-dotenv>=1.0.0
motor>=3.3.0
